+++ CHANGELOG.md
@@ -6,9 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- Method `setCardFormData` and `cardFormData` from payment context.
 
 ## [0.3.0] - 2020-05-12
-## Added
+### Added
 - `setPaymentField` method to update a field of payment 
 - Get data from `OrderForm` and send to `children`